relocate debug client to optional feature
