+++ core/src/lib.rs
@@ -14,6 +14,8 @@ extern crate core;
 mod abstractions;
 #[cfg(feature = "ephemeral-server")]
 pub mod ephemeral_server;
+#[cfg(feature = "debug-plugin")]
+pub mod debug_client;
 mod internal_flags;
 mod pollers;
 mod protosext;
